refactor: extract query-parse crate from query (Part 2 of #670)
Move query_parse/, sdl_parse/, schema_gen/, and select_convert.rs
(~10K lines, 31 files) from the query crate into a new query-parse crate.

Also extract CollectionProvider trait and StaticCollectionProvider from
query::fetcher into query-types to break the circular dependency between
query-parse and query.

The query crate re-exports all modules via extern crate aliases so
downstream imports continue to work unchanged.
