+++ Cargo.toml
@@ -9,6 +9,7 @@ members = [
     "crates/schema",
     "crates/query",
     "crates/query-types",
+    "crates/query-parse",
     "crates/document",
     "crates/p2p",
     "crates/crypto",
