+++ crates/query-parse/Cargo.toml
@@ -1,35 +1,39 @@
 [package]
-name = "query-types"
+name = "query-parse"
 version.workspace = true
 authors.workspace = true
 edition.workspace = true
 rust-version.workspace = true
 license.workspace = true
 repository.workspace = true
-description = "Shared type definitions for the DefraDB query engine"
+description = "GraphQL/SDL parsing and schema generation for DefraDB"
 
 [dependencies]
 # Serialization
-serde.workspace = true
 serde_json.workspace = true
-
-# Error handling
-thiserror.workspace = true
-
-# Async
-async-trait.workspace = true
 
 # Logging
 tracing.workspace = true
 
+# Content addressing (for collection version IDs)
+cid.workspace = true
+sha2.workspace = true
+
+# GraphQL
+graphql-parser.workspace = true
+
 # Time
 chrono = { version = "0.4", features = ["serde"] }
 
+# Regex (SDL preprocessing)
+regex = "1"
+
+# Async
+async-trait.workspace = true
+
 # Internal crates
+query-types = { path = "../query-types" }
 schema = { path = "../schema" }
-storage = { path = "../storage" }
-document = { path = "../document" }
-defra-core = { path = "../defra-core" }
 
 [dev-dependencies]
 tokio = { workspace = true, features = ["rt-multi-thread", "macros"] }
