+++ crates/query-parse/src/query_parse/types.rs
@@ -1,10 +1,9 @@
 //! Types for parsed GraphQL operations
 
-use crate::mapper::{Mutation, Select};
+use query_types::mapper::{Mutation, Select};
 
 /// Type of explain output requested.
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Default)]
-#[non_exhaustive]
 pub enum ExplainType {
     /// Simple explanation showing query plan structure without execution.
     #[default]
@@ -29,7 +28,6 @@ impl ExplainType {
 
 /// Result of parsing a GraphQL request.
 #[derive(Debug)]
-#[non_exhaustive]
 pub enum ParsedOperation {
     /// Query operations (SELECT)
     Query {
