+++ crates/query-parse/src/sdl_parse/parser.rs
@@ -3,7 +3,7 @@
 //! Parses GraphQL Schema Definition Language (SDL) into DefraDB CollectionVersion schemas.
 //! Designed for compatibility with Go DefraDB's SDL parsing behavior.
 
-use crate::error::{QueryError, Result};
+use query_types::error::{QueryError, Result};
 use graphql_parser::schema::{Definition, Document, InterfaceType, ObjectType, TypeDefinition};
 use schema::CollectionVersion;
 use std::collections::HashMap;
