+++ crates/query-types/src/lib.rs
@@ -3,6 +3,7 @@
 //! This crate contains the type vocabulary used across parsing, planning,
 //! and execution: Filter, Select, Mutation, DocumentMapping, Doc, QueryError, etc.
 
+pub mod collection_provider;
 pub mod doc;
 pub mod document;
 pub mod error;
@@ -10,6 +11,7 @@ pub mod json_convert;
 pub mod mapper;
 
 // Re-export primary types
+pub use collection_provider::{CollectionProvider, StaticCollectionProvider};
 pub use doc::{Doc, DocFields, DocStatus};
 pub use document::{DocumentMapping, RenderKey};
 pub use error::{QueryError, Result, TransactionError};
